app/requirements.txt
langchain==0.2.12
langchain-community==0.2.11
sentence-transformers==3.0.1
pypdf==4.3.1
faiss-cpu==1.8.0.post1
langchain-huggingface==0.0.3
langchain-openai==0.1.20
langchainhub==0.1.20
fastapi==0.112.0
cryptography==43.0.0
pytest-asyncio==0.23.8
uvicorn==0.30.5
pydantic-settings==2.4.0
aioredis==2.0.1
celery==5.4.0
msgpack==1.0.8
loguru==0.7.2
pyttsx3==2.90
redisearch==2.1.1